+++ packages/blocks-editor/src/Editor/BlocksEditor.tsx
@@ -18,7 +18,7 @@ import {HashtagPlugin} from '@lexical/react/LexicalHashtagPlugin';
 import {HistoryPlugin} from '@lexical/react/LexicalHistoryPlugin';
 import {LinkPlugin} from '@lexical/react/LexicalLinkPlugin';
 import {ListPlugin} from '@lexical/react/LexicalListPlugin';
-import {EditorState, LexicalEditor} from 'lexical';
+import {$getRoot, EditorState, LexicalEditor} from 'lexical';
 import HorizontalRulePlugin from '../Lexical/Plugins/HorizontalRulePlugin';
 import TwitterPlugin from '../Lexical/Plugins/TwitterPlugin';
 import YouTubePlugin from '../Lexical/Plugins/YouTubePlugin';
@@ -28,26 +28,50 @@ import DraggableBlockPlugin from '../Lexical/Plugins/DraggableBlockPlugin';
 import CodeHighlightPlugin from '../Lexical/Plugins/CodeHighlightPlugin';
 import FloatingTextFormatToolbarPlugin from '../Lexical/Plugins/FloatingTextFormatToolbarPlugin';
 import FloatingLinkEditorPlugin from '../Lexical/Plugins/FloatingLinkEditorPlugin';
+import {truncateString} from './Utils';
+import {SuperEditorContentId} from './Constants';
 
 const BlockDragEnabled = false;
 
 type BlocksEditorProps = {
-  onChange: (value: string) => void;
+  onChange: (value: string, preview: string) => void;
   className?: string;
-  children: React.ReactNode;
+  children?: React.ReactNode;
+  previewLength: number;
+  spellcheck?: boolean;
 };
 
 export const BlocksEditor: FunctionComponent<BlocksEditorProps> = ({
   onChange,
   className,
   children,
+  previewLength,
+  spellcheck,
 }) => {
+  const [didIgnoreFirstChange, setDidIgnoreFirstChange] = useState(false);
   const handleChange = useCallback(
     (editorState: EditorState, _editor: LexicalEditor) => {
-      const stringifiedEditorState = JSON.stringify(editorState.toJSON());
-      onChange(stringifiedEditorState);
+      if (!didIgnoreFirstChange) {
+        setDidIgnoreFirstChange(true);
+        return;
+      }
+
+      editorState.read(() => {
+        const childrenNodes = $getRoot().getAllTextNodes().slice(0, 2);
+        let previewText = '';
+        childrenNodes.forEach((node, index) => {
+          previewText += node.getTextContent();
+          if (index !== childrenNodes.length - 1) {
+            previewText += '\n';
+          }
+        });
+        previewText = truncateString(previewText, previewLength);
+
+        const stringifiedEditorState = JSON.stringify(editorState.toJSON());
+        onChange(stringifiedEditorState, previewText);
+      });
     },
-    [onChange],
+    [onChange, didIgnoreFirstChange],
   );
 
   const [floatingAnchorElem, setFloatingAnchorElem] =
@@ -67,7 +91,9 @@ export const BlocksEditor: FunctionComponent<BlocksEditorProps> = ({
           <div id="blocks-editor" className="editor-scroller">
             <div className="editor" ref={onRef}>
               <ContentEditable
+                id={SuperEditorContentId}
                 className={`ContentEditable__root ${className}`}
+                spellCheck={spellcheck}
               />
             </div>
           </div>
@@ -85,7 +111,7 @@ export const BlocksEditor: FunctionComponent<BlocksEditorProps> = ({
         ]}
       />
       <TablePlugin />
-      <OnChangePlugin onChange={handleChange} />
+      <OnChangePlugin onChange={handleChange} ignoreSelectionChange={true} />
       <HistoryPlugin />
       <HorizontalRulePlugin />
       <AutoFocusPlugin />