+++ packages/blocks-editor/src/Editor/Utils.ts
@@ -1,0 +1,7 @@
+export function truncateString(string: string, limit: number) {
+  if (string.length <= limit) {
+    return string;
+  } else {
+    return string.substring(0, limit) + '...';
+  }
+}