+++ packages/desktop/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.23.293](https://github.com/standardnotes/app/compare/@standardnotes/desktop@3.23.292...@standardnotes/desktop@3.23.293) (2022-11-11)
+
+**Note:** Version bump only for package @standardnotes/desktop
+
+## [3.23.292](https://github.com/standardnotes/app/compare/@standardnotes/desktop@3.23.291...@standardnotes/desktop@3.23.292) (2022-11-11)
+
+**Note:** Version bump only for package @standardnotes/desktop
+
+## [3.23.291](https://github.com/standardnotes/app/compare/@standardnotes/desktop@3.23.290...@standardnotes/desktop@3.23.291) (2022-11-10)
+
+**Note:** Version bump only for package @standardnotes/desktop
+
 ## [3.23.290](https://github.com/standardnotes/app/compare/@standardnotes/desktop@3.23.289...@standardnotes/desktop@3.23.290) (2022-11-10)
 
 **Note:** Version bump only for package @standardnotes/desktop