+++ packages/mobile/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.45.42](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.45.41...@standardnotes/mobile@3.45.42) (2022-11-11)
+
+**Note:** Version bump only for package @standardnotes/mobile
+
+## [3.45.41](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.45.40...@standardnotes/mobile@3.45.41) (2022-11-11)
+
+**Note:** Version bump only for package @standardnotes/mobile
+
+## [3.45.40](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.45.39...@standardnotes/mobile@3.45.40) (2022-11-10)
+
+**Note:** Version bump only for package @standardnotes/mobile
+
 ## [3.45.39](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.45.38...@standardnotes/mobile@3.45.39) (2022-11-10)
 
 **Note:** Version bump only for package @standardnotes/mobile