+++ packages/releases/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.224](https://github.com/standardnotes/app/compare/@standardnotes/releases@1.3.223...@standardnotes/releases@1.3.224) (2022-11-11)
+
+**Note:** Version bump only for package @standardnotes/releases
+
+## [1.3.223](https://github.com/standardnotes/app/compare/@standardnotes/releases@1.3.222...@standardnotes/releases@1.3.223) (2022-11-11)
+
+**Note:** Version bump only for package @standardnotes/releases
+
+## [1.3.222](https://github.com/standardnotes/app/compare/@standardnotes/releases@1.3.221...@standardnotes/releases@1.3.222) (2022-11-10)
+
+**Note:** Version bump only for package @standardnotes/releases
+
 ## [1.3.221](https://github.com/standardnotes/app/compare/@standardnotes/releases@1.3.220...@standardnotes/releases@1.3.221) (2022-11-10)
 
 **Note:** Version bump only for package @standardnotes/releases