+++ packages/web/CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.95.1](https://github.com/standardnotes/app/compare/@standardnotes/web@3.95.0...@standardnotes/web@3.95.1) (2022-11-11)
+
+### Bug Fixes
+
+* super improvements ([#1995](https://github.com/standardnotes/app/issues/1995)) ([da6f36f](https://github.com/standardnotes/app/commit/da6f36f34cdee324339d2c923d61191b373513f3))
+
+# [3.95.0](https://github.com/standardnotes/app/compare/@standardnotes/web@3.94.2...@standardnotes/web@3.95.0) (2022-11-11)
+
+### Features
+
+* **dev:** insert current date and time command in super note ([#1994](https://github.com/standardnotes/app/issues/1994)) ([e27ac31](https://github.com/standardnotes/app/commit/e27ac3126ee4fec9e81df8722b6326ba6f38fe5e))
+
+## [3.94.2](https://github.com/standardnotes/app/compare/@standardnotes/web@3.94.1...@standardnotes/web@3.94.2) (2022-11-10)
+
+### Bug Fixes
+
+* super note improvements ([#1991](https://github.com/standardnotes/app/issues/1991)) ([2dbc895](https://github.com/standardnotes/app/commit/2dbc89594ea2759c49bc318d5efd2b3a5ef3b22d))
+
 ## [3.94.1](https://github.com/standardnotes/app/compare/@standardnotes/web@3.94.0...@standardnotes/web@3.94.1) (2022-11-10)
 
 ### Bug Fixes